requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
sympy==1.12
nltk==3.8.1
ragas==0.0.20
sentence-transformers==2.3.1
scipy==1.11.4
python-dotenv==1.0.0
flask-cors==4.0.0
SpeechRecognition==3.10.0
matplotlib==3.8.2
pandas==2.1.3 